+++ ksa_compliance/ksa_compliance/doctype/zatca_business_settings/zatca_business_settings.js
@@ -12,9 +12,9 @@ frappe.ui.form.on("ZATCA Business Settings", {
         if (!frm.is_new() && frm.doc.status === "Revoked") {
             add_create_business_settings_button(frm)
         }
-      
+
         frm.add_custom_button(__("Submit Feedback"), () => {
-            ksa_compliance.feedback_dialog.show_feedback_dialog(__("Submit Feedback"));
+            ksa_compliance.feedback_dialog.show_feedback_dialog(__("Submit Feedback"), frm.doc.company);
         });
     },
     company: function (frm) {
@@ -174,7 +174,7 @@ frappe.ui.form.on("ZATCA Business Settings", {
             });
 
             if (result.message) {
-                ksa_compliance.feedback_dialog.show_feedback_dialog(__("KSA Compliance Feedback"), true);
+                ksa_compliance.feedback_dialog.show_feedback_dialog(__("KSA Compliance Feedback"), frm.doc.company, true);
             }
         });
     },